requirements.txt
# Thư viện cho các node xử lý file
pandas
requests
openpyxl
python-docx

# Thư viện cho hệ thống đăng nhập
aiohttp_session
bcrypt
Jinja2